[#7] feat : add the day components
